requirements.txt
streamlit>=1.29.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.24.3
spacy>=3.7.2
pdfplumber>=0.10.3
python-docx>=1.1.0
plotly>=5.17.0
matplotlib>=3.8.2
textstat>=0.7.3
nltk>=3.8.1
tqdm>=4.66.0
